+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
  * Created `Add-GitItem` function for promoting new, untracked, and modified files/directories to the Git staging area so they can be committed. Implements the `git add` command.
  * Created `Get-GitRepository` function for getting an object representing a repository.
  * Created `Save-GitChange` function for commiting changes to a repository. Implements the `git commit` command.
+ * Created `Find-GitRepository` function for searching a directory and its parents (i.e. up it tree) for a Git repository.
+
  
 # 0.0.0 (31 August 2016)