+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
  * Created `Save-GitChange` function for commiting changes to a repository. Implements the `git commit` command.
  * Created `Find-GitRepository` function for searching a directory and its parents (i.e. up it tree) for a Git repository.
  * Created `New-GitRepository` function for creating new Git repositories.
+ * Created `Get-GitRepositorStatus` function for getting the state of a repository's working directory and any items staged for the next commit.
 
  
 # 0.0.0 (31 August 2016)