+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-# 0.1.0
+# 0.1.0 (9 September 2016)
 
  * Created `Add-GitItem` function for promoting new, untracked, and modified files/directories to the Git staging area so they can be committed. Implements the `git add` command.
  * Created `Get-GitRepository` function for getting an object representing a repository.